pyproject.toml
[tool.poetry]
name = "fyp"
version = "0.1.0"
description = ""
authors = ["R00231349@mymtu.ie"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
psycopg2-binary = "*"
pandas = "*"
requests = "*"
openai = "*"
python-dotenv = "*"
pyarrow = "*"
datasets = "*"
sacrebleu = "*"
rouge = "*"
streamlit = "*"
matplotlib = "^3.10.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
